chore: type tweaks

+++ api/models/user_create.go
@@ -4,15 +4,15 @@ import "encoding/json"
 
 type UserCreate struct {
 	Proxies                UserProxies                `json:"proxies"`
-	Expire                 uint64                     `json:"expire,omitempty"`
-	DataLimit              uint64                     `json:"data_limit,omitempty"`
+	Expire                 int                        `json:"expire,omitempty"`
+	DataLimit              int                        `json:"data_limit,omitempty"`
 	DataLimitResetStrategy UserDataLimitResetStrategy `json:"data_limit_reset_strategy,omitempty"`
 	Inbounds               UserInbounds               `json:"inbounds,omitempty"`
 	Note                   string                     `json:"note,omitempty"`
 	SubUpdatedAt           *Time                      `json:"sub_updated_at,omitempty"`
 	SubLastUserAgent       string                     `json:"sub_last_user_agent,omitempty"`
 	OnlineAt               *Time                      `json:"online_at,omitempty"`
-	OnHoldExpireDuration   uint64                     `json:"on_hold_expire_duration,omitempty"`
+	OnHoldExpireDuration   int                        `json:"on_hold_expire_duration,omitempty"`
 	OnHoldTimeout          *Time                      `json:"on_hold_timeout,omitempty"`
 	Username               string                     `json:"username,omitempty"`
 	Status                 UserCreateStatus           `json:"status,omitempty"`
